pyproject.toml
[tool.poetry]
name = "qq-chat-history"
version = "0.2.3"
description = "A tool to extract QQ chat history."
authors = ["kifuan <kifuan@foxmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "qq_chat_history"}]
repository = "https://github.com/kifuan/qq-chat-history"


[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
ujson = "^5.7.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
types-ujson = "^5.7.0.0"
mypy = "^0.991"
flake8 = "^6.0.0"


[tool.poetry.scripts]
qq-chat-history = "qq_chat_history.cli:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
